+++ src/Core/Content/Flow/Dispatching/Action/ListrakSendMailAction.php
@@ -2,9 +2,9 @@
 
 namespace Listrak\Core\Content\Flow\Dispatching\Action;
 
-use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Exception;
 use Dotdigital\Flow\Core\Framework\Event\ListrakMailAware;
+use Listrak\Event\ListrakTemplateCustomizationEvent;
 use Listrak\Service\DataMappingService;
 use Listrak\Service\ListrakApiService;
 use Psr\Log\LoggerInterface;
@@ -12,16 +12,14 @@ use Shopware\Core\Content\Flow\Dispatching\Action\FlowAction;
 use Shopware\Core\Content\Flow\Dispatching\Action\FlowMailVariables;
 use Shopware\Core\Content\Flow\Dispatching\DelayableAction;
 use Shopware\Core\Content\Flow\Dispatching\StorableFlow;
-use Shopware\Core\Content\Mail\Service\AbstractMailService;
 use Shopware\Core\Content\MailTemplate\Exception\MailEventConfigurationException;
 use Shopware\Core\Content\MailTemplate\Exception\SalesChannelNotFoundException;
-use Shopware\Core\Framework\Adapter\Translation\AbstractTranslator;
+use Shopware\Core\Framework\Adapter\Twig\StringTemplateRenderer;
 use Shopware\Core\Framework\DataAbstractionLayer\Exception\InconsistentCriteriaIdsException;
 use Shopware\Core\Framework\Event\EventData\MailRecipientStruct;
 use Shopware\Core\Framework\Event\MailAware;
-use Shopware\Core\System\Locale\LanguageLocaleCodeProvider;
 use Shopware\Core\System\SalesChannel\SalesChannelContext;
-
+use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
 
 class ListrakSendMailAction extends FlowAction implements DelayableAction
 {
@@ -36,7 +34,9 @@ class ListrakSendMailAction extends FlowAction implements DelayableAction
     public function __construct(
         private readonly ListrakApiService $listrakApiService,
         private readonly DataMappingService $dataMappingService,
-        private readonly LoggerInterface $logger
+        private readonly StringTemplateRenderer $templateRenderer,
+        private readonly EventDispatcherInterface $eventDispatcher,
+        private LoggerInterface $logger
     ) {
     }
 
@@ -50,7 +50,7 @@ class ListrakSendMailAction extends FlowAction implements DelayableAction
      */
     public function requirements(): array
     {
-        return [ListrakMailAware::class,MailAware::class];
+        return [ListrakMailAware::class, MailAware::class];
     }
 
     /**
@@ -60,21 +60,25 @@ class ListrakSendMailAction extends FlowAction implements DelayableAction
      */
     public function handleFlow(StorableFlow $flow): void
     {
-
         if (!$flow->hasData(MailAware::MAIL_STRUCT) || !$flow->hasData(MailAware::SALES_CHANNEL_ID)) {
             throw new MailEventConfigurationException('Not have data from MailAware', $flow::class);
         }
 
         $eventConfig = $flow->getConfig();
         if (empty($eventConfig['recipient'])) {
-            throw new MailEventConfigurationException('The recipient value in the flow action configuration is missing.', $flow::class);
+            throw new MailEventConfigurationException(
+                'The recipient value in the flow action configuration is missing.',
+                $flow::class
+            );
         }
         if (empty($eventConfig['transactionalMessageId'])) {
-            throw new MailEventConfigurationException('The transactional message ID value in the flow action configuration is missing.', $flow::class);
+            throw new MailEventConfigurationException(
+                'The transactional message ID value in the flow action configuration is missing.',
+                $flow::class
+            );
         }
-
         $transactionalMessageId = $eventConfig['transactionalMessageId'];
-        $profileFields = $eventConfig['profileFields'] ?? [];
+        $profileFields = $eventConfig['profileFields'] ?? '';
 
         /** @var MailRecipientStruct $mailStruct */
         $mailStruct = $flow->getData(MailAware::MAIL_STRUCT);
@@ -89,21 +93,33 @@ class ListrakSendMailAction extends FlowAction implements DelayableAction
             return;
         }
 
+        $templateData = [
+            'eventName' => $flow->getName(),
+            ...$flow->data(),
+        ];
+
         $context = $flow->getContext();
         /** @var SalesChannelContext $channelContext */
+        foreach ($profileFields as $key => $value) {
+            $profileFields[$key] = $this->templateRenderer->render($value, $templateData, $context);
+        }
 
-        $data  = $this->dataMappingService->mapTransactionalMessageData($recipients,$profileFields);
-        $this->listrakApiService->sendTransactionalMessage($transactionalMessageId,$data, $context);
+        $fields = $this->dataMappingService->mapTemplateVariables($profileFields) ?? [];
+        $event = new ListrakTemplateCustomizationEvent($fields, $flow);
+        $this->eventDispatcher->dispatch($event, ListrakTemplateCustomizationEvent::NAME);
+        $modifiedPayload = $event->getPayload();
+        $data = $this->dataMappingService->mapTransactionalMessageData($recipients, $modifiedPayload);
+        $this->listrakApiService->sendTransactionalMessage($transactionalMessageId, $data, $context);
     }
-
 
     /**
      * @param array<string, mixed> $recipients
      * @param array<string, string> $mailStructRecipients
      * @param array<int|string, mixed> $contactFormData
      *
-     * @return array<int|string, string>
      * @throws Exception
+     *
+     * @return array<int|string, string>
      */
     private function getRecipients(array $recipients, $mailStructRecipients, array $contactFormData): array
     {